refactor : 조회 로직 수정, admin과 분리
