+++ inquiry/src/main/java/com/kernel/inquiry/repository/CustomInquiryRepositoryImpl.java
@@ -1,14 +1,10 @@
 package com.kernel.inquiry.repository;
 
-import com.kernel.global.common.enums.UserRole;
-import com.kernel.inquiry.common.exception.CustomNoDataFoundException;
 import com.kernel.inquiry.domain.entity.QInquiry;
-import com.kernel.inquiry.domain.info.InquirySummaryInfo;
+import com.kernel.inquiry.service.info.InquirySummaryInfo;
 import com.kernel.inquiry.service.dto.request.InquirySearchReqDTO;
-
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.JPQLQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -20,8 +16,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-import static com.kernel.inquiry.domain.entity.QInquiry.inquiry;
-
 @Repository
 @RequiredArgsConstructor
 public class CustomInquiryRepositoryImpl implements CustomInquiryRepository {
@@ -32,13 +26,13 @@ public class CustomInquiryRepositoryImpl implements CustomInquiryRepository {
     /**
      * Inquiry 검색 및 페이징 처리
      *
-     * @param request   검색 요청 DTO
-     * @param authorId  작성자 ID
+     * @param searchReqDTO   검색 요청 DTO
+     * @param userId  작성자 ID
      * @param pageable  페이지 정보
      * @return 페이징된 Inquiry 결과
      */
     @Override
-    public Page<InquirySummaryInfo> searchInquiriesWithPagination(InquirySearchReqDTO request, Long authorId, Boolean isAdmin, Pageable pageable) {
+    public Page<InquirySummaryInfo> searchInquiriesWithPagination(InquirySearchReqDTO searchReqDTO, Long userId, Pageable pageable) {
 
         // 전체 개수 조회
         long total = Optional.ofNullable(
@@ -46,41 +40,36 @@ public class CustomInquiryRepositoryImpl implements CustomInquiryRepository {
                         .select(inquiry.count())
                         .from(inquiry)
                         .where(
-                                authorEq(authorId, isAdmin),
-                                authorRoleEq(request.getAuthorRole(), isAdmin),
-                                notDeleted(),
-                                createdAtGoe(request.getFromCreatedAt()),
-                                createdAtLoe(request.getToCreatedAt()),
-                                titleContains(request.getTitleKeyword()),
-                                contentContains(request.getContentKeyword()),
-                                replyStatusCond(request.getReplyStatus(), inquiry)
+                                inquiry.authorId.eq(userId),
+                                inquiry.isDeleted.isFalse(),
+                                createdAtGoe(searchReqDTO.getFromCreatedAt()),
+                                createdAtLoe(searchReqDTO.getToCreatedAt()),
+                                titleContains(searchReqDTO.getTitleKeyword()),
+                                contentContains(searchReqDTO.getContentKeyword()),
+                                replyStatusCond(searchReqDTO.getReplyStatus(), inquiry)
                         )
                         .fetchOne()
         ).orElse(0L);
-
-        // 조회 개수가 0인 경우 예외 처리
-        if (total == 0) {
-            throw new CustomNoDataFoundException("조건에 맞는 문의사항이 없습니다.");
-        }
 
         // 페이지 결과 조회
         List<InquirySummaryInfo> results = jpaQueryFactory
                 .select(Projections.fields(InquirySummaryInfo.class,
                         inquiry.inquiryId,
+                        inquiry.categoryName,
                         inquiry.title,
-                        inquiry.content,
                         inquiry.createdAt,
-                        inquiry.isReplied
+                        inquiry.isReplied,
+                        inquiry.authorType
                 ))
+                .from(inquiry)
                 .where(
-                        authorEq(authorId, isAdmin),
-                        authorRoleEq(request.getAuthorRole(), isAdmin),
-                        notDeleted(),
-                        createdAtGoe(request.getFromCreatedAt()),
-                        createdAtLoe(request.getToCreatedAt()),
-                        titleContains(request.getTitleKeyword()),
-                        contentContains(request.getContentKeyword()),
-                        replyStatusCond(request.getReplyStatus(), inquiry)
+                        inquiry.authorId.eq(userId),
+                        inquiry.isDeleted.isFalse(),
+                        createdAtGoe(searchReqDTO.getFromCreatedAt()),
+                        createdAtLoe(searchReqDTO.getToCreatedAt()),
+                        titleContains(searchReqDTO.getTitleKeyword()),
+                        contentContains(searchReqDTO.getContentKeyword()),
+                        replyStatusCond(searchReqDTO.getReplyStatus(), inquiry)
                 )
                 .orderBy(inquiry.inquiryId.desc()) // 최신순 정렬
                 .offset(pageable.getOffset())
@@ -88,55 +77,6 @@ public class CustomInquiryRepositoryImpl implements CustomInquiryRepository {
                 .fetch();
 
         return new PageImpl<>(results, pageable, total);
-    }
-
-    /**
-     * 작성자 ID에 따라 Inquiry를 조회하는 조건식
-     *
-     * @param authorId 작성자 ID
-     * @param isAdmin  관리자 여부
-     * @return 작성자 ID에 따른 Inquiry 조건식
-     */
-    private BooleanExpression authorEq(Long authorId, Boolean isAdmin) {
-        // 관리자일 경우 작성자 ID가 null인 경우 모든 Inquiry를 조회
-        if (isAdmin != null && isAdmin) {
-            return authorId == null ? null : inquiry.authorId.eq(authorId);
-        }
-        // 일반 사용자의 경우 작성자 ID가 null이 아니어야 함
-        if (authorId == null) {
-            throw new IllegalArgumentException("작성자 ID는 필수입니다.");
-        }
-        return inquiry.authorId.eq(authorId);
-    }
-
-    private BooleanExpression authorRoleEq(String authorRole, Boolean isAdmin) {
-        // 관리자가 아닐 경우 null
-        if (isAdmin != null && !isAdmin) {
-            return Expressions.asBoolean(false);
-        }
-
-        // 관리자가 문의를 조회할 때 작성자 유형이 지정되어야 함
-        if (isAdmin != null && isAdmin && (authorRole == null || authorRole.isEmpty())) {
-            throw new IllegalArgumentException("관리자가 문의를 조회할 때 조회할 작성자 유형을 지정해야합니다.");
-        }
-
-        // 작성자 유형에 따라 조건식 생성
-        if (UserRole.CUSTOMER.name().equals(authorRole)) {
-            return inquiry.authorRole.eq(UserRole.CUSTOMER);
-        } else if (UserRole.MANAGER.name().equals(authorRole)) {
-            return inquiry.authorRole.eq(UserRole.MANAGER);
-        } else {
-            throw new IllegalArgumentException("유효하지 않은 작성자 유형입니다.");
-        }
-    }
-
-    /**
-     * 삭제되지 않은 Inquiry를 조회하는 조건식
-     *
-     * @return 삭제되지 않은 Inquiry 조건식
-     */
-    private BooleanExpression notDeleted() {
-        return inquiry.isDeleted.isFalse();
     }
 
     /**
